fix: Prevent arithmetic overflow and state bugs across widgets and render
- Layout solver: use saturating_add for Max constraint checks (lib.rs:355,364)
- Presenter: use saturating_add for cursor_x tracking (presenter.rs:250)
- Frame/Drawing: widen to u32 for wide-char fit check to prevent u16 overflow
- Frame: reset cursor_position in clear() to prevent stale cursor bleeding
- TextInput: reset scroll_cells in set_value() to prevent stale scroll state
- TextInput: use actual mask char width instead of hardcoded 1
- Pretty: fix row_idx u16 truncation that could bypass height guard
- Pretty: use saturating_add for y position calculation
- Virtualized: clamp scroll to len-visible_count instead of len-1

+++ crates/ftui-render/src/drawing.rs
@@ -211,7 +211,7 @@ impl Draw for Buffer {
             }
 
             // Don't start a wide char if it won't fit
-            if cx + width as u16 > max_x {
+            if cx as u32 + width as u32 > max_x as u32 {
                 break;
             }
